Crate solana_stake_interface

The Stake program interface.

Modules

config

Config for staking.

error

instruction

program

stake_flags

stake_history

A type to hold data for the StakeHistory sysvar.

state

tools

Utility functions

Constants

MINIMUM_DELINQUENT_EPOCHS_FOR_DEACTIVATION

The minimum number of epochs before stake account that is delegated to a delinquent vote account may be unstaked with StakeInstruction::DeactivateDelinquent