solana_stake_interface::state

Struct Lockup

pub struct Lockup {
    pub unix_timestamp: UnixTimestamp,
    pub epoch: Epoch,
    pub custodian: Pubkey,
}

Fields
unix_timestamp: UnixTimestamp

UnixTimestamp at which this stake will allow withdrawal, unless the transaction is signed by the custodian

epoch: Epoch

epoch height at which this stake will allow withdrawal, unless the transaction is signed by the custodian

custodian: Pubkey

custodian signature on a transaction exempts the operation from lockup constraints

Implementations

impl Lockup

pub fn is_in_force(&self, clock: &Clock, custodian: Option<&Pubkey>) -> bool